chore: remove react-hook-form and its related files
