--- frontend/src/components/settings/dynamic-form.tsx
@@ -1,77 +1,0 @@
-import { AlertCircle } from 'lucide-react'
-import { useWatch } from 'react-hook-form'
-
-import { ControlledFieldRenderer } from '@/components/form/controlled-field-renderer'
-import { Alert, AlertDescription } from '@/components/ui/alert'
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
-import type { GroupSchema } from '@/types/settings/group-schema'
-
-import { GroupTitle } from '../layout/group-title'
-import { SettingsSchemaSkeleton } from '../ui/skeleton'
-
-export type SettingsDynamicFormProps = {
-  groupSchema: GroupSchema | null | undefined
-  isInitialLoading?: boolean
-  isRefreshing?: boolean
-}
-
-export const SettingsDynamicForm = ({ groupSchema, isInitialLoading, isRefreshing }: SettingsDynamicFormProps) => {
-  const formValues = useWatch()
-
-  if (isInitialLoading) {
-    return <SettingsSchemaSkeleton />
-  }
-
-  if (!groupSchema) {
-    return (
-      <Alert>
-        <AlertCircle className="h-4 w-4" />
-        <AlertDescription>No settings schema available.</AlertDescription>
-      </Alert>
-    )
-  }
-
-  return (
-    <div className="flex flex-col gap-y-4">
-      <GroupTitle label={groupSchema?.label} icon={groupSchema?.icon} />
-
-      {groupSchema?.sections?.map((section, idx) => {
-        return (
-          <Card key={`${section?.id}-${idx}`} className="flex flex-col gap-y-4">
-            <CardHeader>
-              <CardHeader>
-                <CardTitle>{section?.title}</CardTitle>
-
-                <CardDescription>{section?.subtitle}</CardDescription>
-              </CardHeader>
-            </CardHeader>
-
-            <CardContent className="flex flex-col gap-y-8 max-w-3xl">
-              {section?.fields?.map((field) => {
-                const shouldShow = Object.entries(field?.showIf ?? {}).every(([key, expectedValue]) => {
-                  return formValues[key] === expectedValue
-                })
-
-                const shouldHide = Object.entries(field?.hideIf ?? {}).some(([key, expectedValue]) => {
-                  return formValues[key] === expectedValue
-                })
-
-                if (!shouldShow || shouldHide || Boolean(field?.hidden)) {
-                  return null
-                }
-
-                return (
-                  <ControlledFieldRenderer
-                    isLoading={isRefreshing}
-                    key={`section-${section?.id}-field-${field?.key}`}
-                    schema={field}
-                  />
-                )
-              })}
-            </CardContent>
-          </Card>
-        )
-      })}
-    </div>
-  )
-}
